move entry_is_empty function to util-like file
